+++ crates/data/src/python/mod.rs
@@ -16,7 +16,6 @@
 //! Python bindings from [PyO3](https://pyo3.rs).
 
 pub mod config;
-pub mod option_chain_manager;
 
 use pyo3::prelude::*;
 
@@ -28,6 +27,5 @@ use pyo3::prelude::*;
 #[pymodule]
 pub fn data(_: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<crate::engine::config::DataEngineConfig>()?;
-    m.add_class::<option_chain_manager::PyOptionChainManager>()?;
     Ok(())
 }